Enum re_analytics::Property

pub enum Property {
    Bool(bool),
    Integer(i64),
    Float(f64),
    String(String),
}

Variants

Bool(bool)

Integer(i64)

Float(f64)

String(String)

Implementations

impl Property

pub fn hashed(&self) -> Property

Returns a new string property that is a hex representation of the hashed sum of the current property.