re_analytics
Module posthog

Structs

PostHogBatch

PostHogCaptureEvent

PostHogIdentifyEvent

Enums

PostHogEvent

Constants

PUBLIC_POSTHOG_PROJECT_KEY

The “public” API key can be obtained at https://eu.posthog.com/project/settings#project-api-key. Make sure you are logged in to the right organization and have the correct project open. Unfortunately that stuff is client-side routed, and there’s no way to link directly to the right place.