Trait re_analytics::Properties

pub trait Properties: Sized {
    // Provided method
    fn serialize(self, event: &mut AnalyticsEvent) { ... }
}

Trait representing the properties of an analytics event.

This is separate from Event to facilitate code re-use.

For example, [re_build_info::BuildInfo] has an implementation of this trait, so that any event which wants to include build info in its properties may include that struct in its own definition, and then call build_info.serialize in its own serialize implementation.

Provided Methods

fn serialize(self, event: &mut AnalyticsEvent)

Object Safety

This trait is not object safe.