Crate re_analytics

Rerun’s analytics SDK.

We never collect any personal identifiable information. You can always opt-out with rerun analytics disable.

No analytics will be collected the first time you start the Rerun viewer, giving you an opportunity to opt-out first if you wish.

All the data we collect can be found in event.

Modules

cli

event

All analytics events collected by the Rerun viewer are defined in this file.

native 🔒

posthog 🔒

Structs

Analytics

AnalyticsEvent

Config

Enums

AnalyticsError

ConfigError

EventKind

Property

Constants

DISCLAIMER 🔒

Traits

Event

An analytics event.

Properties

Trait representing the properties of an analytics event.

Functions

load_config 🔒