re_analytics
Module event

All analytics events collected by the Rerun viewer are defined in this file.

Analytics can be completely disabled with rerun analytics disable, or by compiling rerun without the analytics feature flag.

All collected analytics data is anonymized, stripping all personal identifiable information as well as information about user data. Read more about our analytics policy at https://github.com/rerun-io/rerun/tree/main/crates/utils/re_analytics.

Structs

CrashPanic

Records a crash caused by a panic.

CrashSignal

Identify

Holds information about the user’s environment.

OpenRecording

Sent when a new recording is opened.

ServeWasm

Sent when a Wasm file is served.

StoreInfo

Basic information about a recording’s chunk store.

ViewerRuntimeInformation

Some sparse information about the runtime environment the viewer is running in.

ViewerStarted

Sent when the viewer is first started.

Enums

Id

Constants

RERUN_DOMAINS 🔒

Functions

add_sanitized_url_properties 🔒

extract_root_domain 🔒

Given a URL, extract the root domain.